Some (big) modifications to paths.
First, validation is painfully slow, so I've added an option to validate less often and to validate on smaller sets.

Second, the initial h0 for the conditonal was all wrong and too complex. It is now the output of a network starting at h (the latent variable of the sbn)


Former-commit-id: fd0aa28ed6b70048f703dcab80b8886e77729645 [formerly bfdce49bbeafbd7eee4d9ec1ec28bc8759276e62] [formerly 10fd13fadad199a7eee2044585d60c3ffeb3cccf [formerly ec7636ef1817c3921c0d561c9f1bd8ba6a56629b]]
Former-commit-id: 945b44deec9d7b6dbba4b242cc78f121c0b07e87 [formerly 604af96c92fbd1f8718c9adad25011ae4d794fd3]
Former-commit-id: 2ee8a7c0de3b5cdfe53f21622c561159fbc0d546
Former-commit-id: dacc78bd473ae664364181320722c82a830d7eea